minimal working piece

+++ src/js/ui/table/handleTableButtons.js
@@ -1,9 +1,5 @@
-import {
-	setCurrentlyBeignEdited,
-	clearCurrentlyBeignEdited,
-	deleteEntry,
-	updateEntry,
-} from '../../store/actionCreators/entry.js';
+import { store } from '../../store';
+import { setCurrentlyBeignEdited, deleteEntry } from '../../store/actionCreators/entry.js';
 
 import openEntryModel from '../Model/openEntryModel.js';
 import setEntryModelValues from '../Model/setEntryModelValues.js';
@@ -13,7 +9,8 @@ export default () => {
 	const actionsButtons = document.querySelectorAll('.dashboard table tbody .action-buttons .button');
 
 	actionsButtons.forEach((actionButton) => {
-		actionButton.addEventListener('click', function () {
+		actionButton.addEventListener('click', function (e) {
+			e.preventDefault();
 			const entryId = this.parentNode.getAttribute('id');
 			const { buttonType } = this.dataset;
 
@@ -26,7 +23,10 @@ export default () => {
 						openEntryModel();
 						break;
 					case 'delete':
-						deleteEntry(entryId);
+						var rconfirm = confirm(`Do you really want to delete`);
+						if (rconfirm) {
+							return deleteEntry(entryId);
+						}
 						break;
 					default:
 						break;
